\name{getPages}
\alias{getPages}
\title{Get a list of XML nodes for the pages in the PDF document.}
\description{
}
\usage{
getPages(doc)
}
\arguments{
  \item{doc}{
}
}
\details{
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
}
\author{
  Duncan Temple Lang
}
\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{IO}
